// src/app/components/vehicle/vehicle.enums.ts
export enum VEHICLE_SERVICESTATUS_ENUM {
  Active = 'Active',
  OutOfServices = 'OutOfServices',
  Decomissioned = 'Decomissioned',
}
export enum VEHICLE_OPERATIONALSTATUS_ENUM {
  OnService = 'OnService',
  Parked = 'Parked',
}
export enum VEHICLE_COLOR_ENUM {
  Black = 'Black',
  Silver = 'Silver',
  Gray = 'Gray',
  Blue = 'Blue',
  Red = 'Red',
  Yellow = 'Yellow',
  Green = 'Green',
  White = 'White',
}